% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_epinow.R
\name{df_epinow}
\alias{df_epinow}
\title{Manipulate a dataset to use in an Epinow2 analysis}
\usage{
df_epinow(x, cases, date, regions, period = 14, regional = FALSE)
}
\arguments{
\item{x}{a vertical dataset typically used in ecological analyses.}

\item{cases}{the vector of cases in the provided dataset.}

\item{date}{the vector of dates in the format yyyy-mm-dd in the provided dataset.}

\item{regions}{if regional = TRUE, the vector of regions in the provided dataset.}

\item{period}{numeric. How many date periods should be left into the dataset?}

\item{regional}{logical. Does the user wants the dataset to be aggregated by regions?}
}
\description{
\code{df_epinow} manipulates a dataset provided by the user to a dataset readily usable for Epinow2 analyses
}
\details{
This function loads a \code{data.frame} object and makes some manipulation of it, in order to make it ready to be run in a Epinow2 local or regional analysis.
}
\examples{
\dontrun{
library("epigraphhub")
con <- egh_connection(auto_connect = TRUE, use_env = TRUE, path_env = ".env")
df_canton <- dbGetQuery(con, "SELECT datum, \"geoRegion\", entries FROM switzerland.foph_cases")
df <- df_epinow(df_canton, cases = "entries", date = "datum", regions = "geoRegion", regional = TRUE, period = 28)
}
}
