% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2_datasets.R
\name{dhis2_datasets}
\alias{dhis2_datasets}
\title{Get all currently available datasets from a DHIS2 server.}
\usage{
dhis2_datasets(api_url, username, password)
}
\arguments{
\item{api_url}{the URL of the DHIS2 api. Usually it is the server URL ending with "/api/"}

\item{username}{the username required to access the DHIS2 server.}

\item{password}{the password required to access the DHIS2 server.}
}
\value{
A list with the unique id and names of each available dataset.
}
\description{
\code{dhis2_datasets} queries a DHIS2 server and returns all available datasets.
}
\section{Warning}{

A Internet connection is needed to use this function.

Watch out for hardcoded login and password.
}

\examples{
\dontrun{
# Demo available at https://play.dhis2.org/demo/
library("r-epigraphhub")
api_url <- "https://play.dhis2.org/demo/api/"
username <- "admin"
password <- "district"
datasets <- dhis2_datasets(api_url = api_url, username = username, password = password)
datasets
}
}
