% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2_import.R
\name{dhis2_import}
\alias{dhis2_import}
\title{Downloads a dataset from a DHIS2 server. Currently works on datasets organized weekly or monthly.}
\usage{
dhis2_import(
  df,
  dataset,
  api_url,
  username,
  password,
  year_initial,
  year_final,
  period = c("weekly", "monthly")
)
}
\arguments{
\item{df}{a \code{dhis2_datasets} object.}

\item{dataset}{string. The name of an available dataset from a DHIS2 server obtained from a \code{dhis2_datasets} object.}

\item{api_url}{the URL of the DHIS2 api. Usually it is the server URL ending with "/api/"}

\item{username}{the username required to access the DHIS2 server.}

\item{password}{the password required to access the DHIS2 server.}

\item{year_initial}{the initial year available in the dataset to import data from.}

\item{year_final}{the last year available in the dataset to import data from.}

\item{period}{either "weekly" or "monthly" periods are supported.}
}
\value{
A dataset containing all available data.
}
\description{
\code{dhis2_import} downloads a dataset from a DHIS2 server. It works on datasets that are organized weekly or monthly.
}
\section{Warning}{

A Internet connection is needed to use this function.

Watch out for hardcoded login and password.
}

\examples{
\dontrun{
# Demo available at https://play.dhis2.org/demo/
library("r-epigraphhub")
api_url <- "https://play.dhis2.org/demo/api/"
username <- "admin"
password <- "district"
datasets <- dhis2_datasets(api_url = api_url, username = username, password = password)
my_dataset <- datasets$displayName[2]
df_dhis2 <- dhis2_import(df = datasets, dataset = my_dataset, api_url = api_url, username = username, password = password, year_initial = 2021, year_final = 2022, period = "monthly")
head(df_dhis2)
}
}
