% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egh_connection.R
\name{egh_connection}
\alias{egh_connection}
\title{Connect to the EpiGraphHub database via SSH tunneling}
\usage{
egh_connection(
  auto_connect = TRUE,
  use_env = TRUE,
  path_env = ".env",
  user,
  password
)
}
\arguments{
\item{auto_connect}{logical. Should the function use a default connection? Automatically calls the session terminal and makes the connection.}

\item{use_env}{logical. Should the function use the parameters configured in a local .env file?}

\item{path_env}{character. If use_env = TRUE, the user should provide a path to the local .env file.}

\item{user, password}{if use_env = FALSE, please provide access parameters to the EpiGraphHub database.}
}
\description{
\code{egh_connection} connects to the EpiGraphHub database using a .env local file or by manually providing information.
}
\details{
This function connects to the EpiGraphHub (https://epigraphhub.org) database via SSH tunneling. User is required to have a SSH key. Default connection parameter can be used or provided by the user. The user can load a local .env file or provide the parameters inside the \code{egh_connection}.
}
\section{Warning}{

A Internet connection is needed to use this function.

Watch out for hardcoded login and password.
}

\examples{
\dontrun{
egh_connection(auto_connect = TRUE, use_env = TRUE, path_env = ".env")
egh_connection(auto_connect = TRUE, use_env = FALSE, user = "user", password = "password")
}
}
