% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egh_spatial.R
\name{egh_spatial}
\alias{egh_spatial}
\title{Function designed to help user explore and visualize their spatial epi analysis in an interactive way.}
\usage{
egh_spatial(x, is_file = FALSE, area, outcome = NULL)
}
\arguments{
\item{x}{a shapefile, geopackage file, a RDS file, a sf object or a sp object containing the spatial polygons for analysis.}

\item{is_file}{is x a spatial polygon file (.shp, .gpkg or .rds)? Mark FALSE for sf or sp objects.}

\item{area}{the variable in x that contains the vector of names of the polygons.}

\item{outcome}{the outcome variable that is included in x to be analyzed.}
}
\description{
\code{egh_spatial} was designed to help user explore and visualize their spatial epi analysis in an interactive way. It allows to visualize neighbourhood contiguity between polygons, explore their spatial data and visualize some common statistical tests for spatial data. This function relies heavily on other spatial analysis functions included in other packages. User can either load a spatial polygon file (.rds, .gpkg. or .shp files) or provide a sf or sp object to the function.
}
\details{
It currently only allows analysis and visualization of static datasets (e.g., it does not allow datasets with multiple repetitions, like in a time series).

The following will soon be implemented: plotting choropleth map of the dependent variable, plotting a choropleth map of Empirical Bayesian Estimates (global and local), returning Moran's I statistic, running a LISA showing scatterplot and map, running Getis-Ord analysis, running a SaTScan analysis.
}
\section{Warning}{

Calculations might take a while in slower computers. When exploring the different spatial contiguity functions, the higher the number of neighbours and distance the slower the computation time can become.
}

\examples{
\dontrun{
library(epigraphhub)
library(readr)
library(sf)

x <- read_rds('shape_AGO_2.rds')

egh_spatial(x = x, area = "NAME_2")
}
}
