% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temperature_data.R
\name{get_temperature_data}
\alias{get_temperature_data}
\title{Wrapper function to download daily temperature data from Google Earth to use on epi analysis}
\usage{
get_temperature_data(
  x,
  id_col = NULL,
  country = NULL,
  admin_lvl = NULL,
  startDate,
  endDate
)
}
\arguments{
\item{x}{an \code{egh_connection} object or a sf object, needed to pull the specified spatial polygon}

\item{id_col}{if providing a sf object, please specify what is the id column that uniquely specifies each of the spatial objects, such as districts, regions, cities, etc.}

\item{country}{a three-letter string specifying the country iso code (ISO 3166-1 alfa-3) pull data from. Must be "1", "2" or "3". Please, check https://gadm.org/ for available countries and administrative levels.}

\item{admin_lvl}{a string specifying the administrative level to pull data from. Must be "1", "2" or "3". Please, check https://gadm.org/ for available countries and administrative levels.}

\item{startDate}{a string containing the starting date in "yyyy-mm-dd" format for the query.}

\item{endDate}{a string containing the ending date in "yyyy-mm-dd" format for the query.}
}
\description{
\code{get_temperature_data} connects to the EpiGraphHub database to pull an spatial polygon from a country and an specific administrative level. Then, it connects to the Google Earth Engine database to pull a dataframe containing daily temperature data for the study area. This function pulls data from MODIS daily Land-Surface Temperature (LST) at 1km grids using the rgee package (MOD11A1 V6.1 product).
}
\details{
To read more about MOD11A1 V6.1 product:
https://lpdaac.usgs.gov/products/mod11a1v061/

To read more about R Google Earth Engine
https://cran.r-project.org/web/packages/rgee/
}
\section{Warning}{

A Internet connection is needed to use this function. Longer time periods and large datasets can take several minutes to fetch data, especially if using a free account. Adjust expectations accordingly.

This functions calls the rgee package. If you never used rgee, please follow the instalation instructions. You will be prompted to create an user account on Google Earth Engine or to connect to an existing one. This is mandatory for rgee to run. Please, go to https://cran.r-project.org/web/packages/rgee/ for more instructions.
}

\examples{
\dontrun{
library(epigraphhub)
country = "AGO"
admin_lvl = "2"
startDate = "2021-01-01"
endDate = "2022-01-01"
con <- egh_connection(auto_connect = TRUE, use_env = TRUE, path_env = ".env")
df <- get_temperature_data(x = con,
                           country = country,
                           admin_lvl = admin_lvl,
                           startDate = startDate,
                           endDate = endDate)
}
}
