% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb_data.R
\name{get_wb_data}
\alias{get_wb_data}
\title{Downloads all available data from one or more datasets from the World Bank database.}
\usage{
get_wb_data(datasets, start_year = 1960, end_year = 2022)
}
\arguments{
\item{datasets}{a \code{wb_data} object or a vector of strings containing the names of the datasets to be downloaded.}

\item{start_year}{starting year to pull data from. Must be 1960 or higher.}

\item{end_year}{last year to pull data from. Must be higher than start_year.}
}
\description{
\code{get_wb_data} generalizes the function \code{WDI::WDI()} to download the whole dataset instead of single indicators. It defaults to a starting date of 1960 up to 2022. It returns a list of data.frames.
}
\section{Warning}{

An internet connection is needed to use this function.
}

\examples{
\dontrun{
library("r-epigraphhub")
datasets <- wb_data()
my_dataset <- datasets[2]
my_data <- get_wb_data(my_dataset)

my_dataset <- "Environment, Social and Governance (ESG) Data"
my_data <- get_wb_data(my_dataset)
}
}
