% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_data.R
\name{wb_data}
\alias{wb_data}
\title{Get all currently available datasets and indicators from World Bank database.}
\usage{
wb_data()
}
\description{
\code{wb_data} pulls all datasets and the respective indicators currently available from World Bank database. It returns an object consisting of a list of datasets and sublists of indicators. This object can be used to download the data using the \code{get_wb_data} function
}
\section{Warning}{

An internet connection is needed to use this function.
}

\examples{
\dontrun{
library("r-epigraphhub")
datasets <- wb_data()
my_dataset <- datasets[2]
my_dataset
}
}
